+++ contracts/native/governance/node_manager/utils.go
@@ -20,11 +20,77 @@ package node_manager
 
 import (
 	"fmt"
+	"github.com/cosmos/cosmos-sdk/x/distribution/types"
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/contracts/native"
 	"github.com/ethereum/go-ethereum/core"
 	"math/big"
 )
+
+var RatioDecimal = new(big.Int).SetUint64(1000000)
+
+// IncreaseValidatorPeriod return the period just ended
+func IncreaseValidatorPeriod(s *native.NativeContract, validator *Validator) (uint64, error) {
+	dec, err := hexutil.Decode(validator.ConsensusPubkey)
+	if err != nil {
+		return 0, fmt.Errorf("IncreaseValidatorPeriod, decode pubkey error: %v", err)
+	}
+
+	// fetch current rewards
+	validatorAccumulatedRewards, err := GetValidatorAccumulatedRewards(s, dec)
+	if err != nil {
+		return 0, fmt.Errorf("IncreaseValidatorPeriod, GetValidatorAccumulatedRewards error: %v", err)
+	}
+
+	// calculate current ratio
+	// mul decimal
+	rewardsD := new(big.Int).Mul(validatorAccumulatedRewards.Rewards, RatioDecimal)
+	ratio := new(big.Int).Div(rewardsD, validator.TotalStake)
+
+	// fetch snapshot rewards for last period
+	validatorSnapshotRewards, err := GetValidatorSnapshotRewards(s, dec, validatorAccumulatedRewards.Period-1)
+	if err != nil {
+		return 0, fmt.Errorf("IncreaseValidatorPeriod, GetValidatorSnapshotRewards error: %v", err)
+	}
+
+	// decrement reference count
+	err = decreaseReferenceCount(s, dec, validatorAccumulatedRewards.Period-1)
+	if err != nil {
+		return 0, fmt.Errorf("IncreaseValidatorPeriod, decreaseReferenceCount error: %v", err)
+	}
+
+	// set new snapshot rewards with reference count of 1
+	newValidatorSnapshotRewards := &ValidatorSnapshotRewards{
+		AccumulatedRewardsRatio: new(big.Int).Add(validatorSnapshotRewards.AccumulatedRewardsRatio, ratio),
+		ReferenceCount:          1,
+	}
+	err = setValidatorSnapshotRewards(s, dec, validatorAccumulatedRewards.Period, newValidatorSnapshotRewards)
+	if err != nil {
+		return 0, fmt.Errorf("IncreaseValidatorPeriod, setValidatorSnapshotRewards error: %v", err)
+	}
+
+	// set accumulate rewards, incrementing period by 1
+	newValidatorAccumulatedRewards := &ValidatorAccumulatedRewards{
+		Rewards: new(big.Int),
+		Period:  validatorAccumulatedRewards.Period + 1,
+	}
+	err = setValidatorAccumulatedRewards(s, dec, newValidatorAccumulatedRewards)
+	if err != nil {
+		return 0, fmt.Errorf("IncreaseValidatorPeriod, setValidatorAccumulatedRewards error: %v", err)
+	}
+
+	return validatorAccumulatedRewards.Period, nil
+}
+
+func withdrawDelegationRewards(s *native.NativeContract, validator *Validator, stakeInfo *StakeInfo) error {
+	// get stake starting info
+	stakeStaringInfo, err := GetStakeStartingInfo(s, stakeInfo.StakeAddress)
+	if err != nil {
+		return fmt.Errorf("withdrawDelegationRewards, GetStakeStartingInfo error: %v", err)
+	}
+
+}
 
 func nativeTransfer(s *native.NativeContract, from, to common.Address, amount *big.Int) error {
 	if amount.Sign() <= 0 {
