+++ contracts/native/utils/params.go
@@ -39,7 +39,6 @@ var (
 	CrossChainManagerContractAddress = common.HexToAddress("0x5747C05FF236F8d18BB21Bc02ecc389deF853cae")
 	SideChainManagerContractAddress  = common.HexToAddress("0x864Ff06eC5fFc75aB6eaf64263308ef5fa7d6637")
 	NodeManagerContractAddress       = common.HexToAddress("0xA4Bf827047a08510722B2d62e668a72FCCFa232C")
-	DistributeContractAddress        = common.HexToAddress("0xA4Bf827047a08510722B2d62e668a72FCCFa2300")
 	RelayerManagerContractAddress    = common.HexToAddress("0xA22f301D7Cb5b50dcA4a015b12EC0cc5f3971412")
 	Neo3StateManagerContractAddress  = common.HexToAddress("0x5E839898821dB2A2F0eC9F8aAE7D7053744DB051")
